fix: dto에서 price type을 big demical에서 int로 변경

+++ src/main/kotlin/com/mss/assignment/domain/coordination/CoordinationService.kt
@@ -52,12 +52,12 @@ class CoordinationService(
             cheapestProduct =
                 ProductWithBrandAndPrice(
                     brandName = cheapestProduct.brand.name,
-                    price = cheapestProduct.price,
+                    price = cheapestProduct.price.toInt(),
                 ),
             mostExpensiveProduct =
                 ProductWithBrandAndPrice(
                     brandName = mostExpensiveProduct.brand.name,
-                    price = mostExpensiveProduct.price,
+                    price = mostExpensiveProduct.price.toInt(),
                 ),
         )
     }
