+++ src/main/kotlin/com/mss/assignment/domain/coordination/response/CheapestCoordinationByBrand.kt
@@ -6,17 +6,17 @@ import java.math.BigDecimal
 data class CheapestCoordinationByBrand(
     val brandName: String,
     val products: List<ProductWithCategoryAndPrice>,
-    val totalPrice: BigDecimal,
+    val totalPrice: Int,
 ) {
     data class ProductWithCategoryAndPrice(
         val categoryName: String,
-        val price: BigDecimal,
+        val price: Int,
     ) {
         companion object {
             fun fromProduct(product: Product): ProductWithCategoryAndPrice {
                 return ProductWithCategoryAndPrice(
                     categoryName = product.category.name,
-                    price = product.price,
+                    price = product.price.toInt(),
                 )
             }
         }
@@ -27,7 +27,7 @@ data class CheapestCoordinationByBrand(
             return CheapestCoordinationByBrand(
                 brandName = product.first().brand.name,
                 products = product.map { ProductWithCategoryAndPrice.fromProduct(it) },
-                totalPrice = product.sumOf { it.price },
+                totalPrice = product.sumOf { it.price }.toInt(),
             )
         }
     }
