+++ src/main/kotlin/com/mss/assignment/dto/ProductDto.kt
@@ -1,19 +1,18 @@
 package com.mss.assignment.dto
 
 import com.mss.assignment.domain.product.Product
-import java.math.BigDecimal
 
 data class ProductDto(
     val categoryName: String,
     val brandName: String,
-    val price: BigDecimal,
+    val price: Int,
 ) {
     companion object {
         fun fromEntity(product: Product): ProductDto {
             return ProductDto(
                 categoryName = product.category.name,
                 brandName = product.brand.name,
-                price = product.price,
+                price = product.price.toInt(),
             )
         }
     }
